[refactor] 諸々整理
