+++ inv_aki_flask/app.py
@@ -1,11 +1,15 @@
+from datetime import datetime
+from hashlib import md5
+
 from flask import Flask
 
-from inv_aki_flask.views import index, main
+from inv_aki_flask.views import index, login, main
 
 app = Flask(__name__)
 app.register_blueprint(index.view)
 app.register_blueprint(main.view)
-app.secret_key = b"random string..."
+app.register_blueprint(login.view)
+app.secret_key = md5(str(datetime.now()).encode("utf-8")).hexdigest()
 
 
 def launch_for_local():
