+++ inv_aki_flask/templates/layout.html
@@ -11,7 +11,7 @@
         {% block content %}{% endblock %}
 
         <div class="footer">
-            {% block footer %}{% endblock %}
+            copyright 2023 tkomota
         </div>
     </body>
 </html>
