+++ inv_aki_flask/views/main.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, redirect, render_template, request, session
+from flask import Blueprint, redirect, render_template, request, session, url_for
 
 view = Blueprint("main", __name__, url_prefix="/main")
 
@@ -19,57 +19,13 @@ def show():
             data=message_data,
         )
     else:
-        return redirect("/main/login")
+        return redirect(url_for("login.show"))
 
 
 @view.route("/", methods=["POST"])
-def form():
+def post():
     global message_data
     msg = request.form.get("comment")
     message_data.append((session["id"], msg))
     message_data = message_data[-25:]
-    return redirect("/main/")
-
-
-@view.route("/login", methods=["GET"])
-def login():
-    return render_template(
-        "login.html",
-        title="Login",
-        err=False,
-        message="IDとパスワードを入力: ",
-        id="",
-    )
-
-
-@view.route("/login", methods=["POST"])
-def login_post():
-    global member_data
-    id = request.form.get("id")
-    pswd = request.form.get("pass")
-
-    if id in member_data:
-        session["login"] = pswd == member_data[id]
-    else:
-        member_data[id] = pswd
-        session["login"] = True
-
-    session["id"] = id
-
-    if session["login"]:
-        return redirect("/main/")
-    else:
-        return render_template(
-            "login.html",
-            title="Login",
-            err=False,
-            message="パスワードが違います",
-            id=id,
-        )
-
-
-@view.route("/logout", methods=["GET"])
-def logout():
-    session.pop("id")
-    session.pop("login")
-    return redirect("/main/login")
+    return redirect(url_for("main.show"))
